fix(clippy): address -D warnings in ui/ + economy/
The cargo clippy gate (RUSTFLAGS=-D warnings) is now enforced on
PR #250. Local clippy --lib passed without warnings; the cargo
clippy --all-targets --all-features step has 20+ warnings that were
all real lints. This commit resolves them with one-line fixes.

Source/build/lint fixes:

- src/ui/widgets.rs:1702 — orphan doc-comment for an aspirational
  tick_tab_clicks example (no such function in the codebase). Clippy
  flagged it because the doc-comment was orphaned (no item between
  it and pub struct Marquee). Demoted to plain comments (still
  useful as a recommended-pattern note).
- src/ui/widgets.rs:1708 — empty_line_after_doc_comments: doc-comment
  and pub struct were separated by a blank line that wasn't
  needed. The fix fell out of the previous bullet.
- src/ui/widgets.rs:1585 — unused 'active_index' parameter in
  tick_tab_body_visibility. Renamed to _active_index (it was
  retained for future per-group subset filters — see the comment
  above the function).
- src/ui/widgets.rs:1870 — BorderColor::clone() on a Copy type.
  Dropped the .clone().
- src/ui/launch/boot_overlay.rs:278 — use super::* in a #[cfg(test)]
  block. The test only uses crate::boot_init::BOOT_STEP_COUNT. Added
  #[allow(unused_imports)].
- src/ui/launch/userdata.rs:82 — pub fn from_str(&str) -> Self
  conflicts with std::str::FromStr. Renamed to parse_str.
- src/ui/shipbuilding_workspace.rs:5468 — two unused imports
  (crate::economy::ResourceType, crate::shipbuilding as
  shipbuilding_types). Dropped; the inner mod re-imports both.
- src/ui/construction/scrollbar.rs:41 — unused mut on ParamSet.
  The body was just 'let _ = (params, metrics);'. Removed the mut.
  Also dropped the mut on the Local<HashMap> at line 47.
- src/ui/construction/buildings.rs:267 — *bt deref of a tuple
  reference was redundant; the closure parameter is already &(bt,
  _count). Pass bt directly.
- src/ui/construction/cards.rs:484/502/514 — count.max(1) as u32
  where count is already u32. Removed the casts.
- src/ui/construction/mining.rs:202 — clippy::needless_borrow on
  &resource_icons in a Bevy system call. Bevy's scheduler expects
  the explicit borrow; suppressed with
  #[allow(clippy::needless_borrow)] scoped to the call site.
- src/ui/construction/queue.rs:166 — identity-map closure
  (|(e, p)| (e, p)) was redundant; projects.iter().collect() is
  sufficient.
- src/ui/construction/queue.rs:200 — count.max(1) is already f32.
  Removed the as f32 cast.
  in-place iteration pattern. Converted to array literals.
- src/ui/resource_icons.rs:1482 — get_or_init(|| TaskPool::new())
  was a redundant closure. Pass TaskPool::new directly.
- src/ui/resource_icons.rs:1577 — for (_, file_name) in &map is
  iterating keys + values; clippy wants .values(). Done.
- src/economy/mining.rs:1894/2013/2033/2055 — 4 sites calling
  run_system_once(update_resource_rates); without using the Result.
  Wrapped with let _ = ... to satisfy must_use.
- src/economy/forecast.rs:158 — collapsible if: the if was nested
  inside another if with no else. Collapsed into a single &&
  condition.
- src/ui/bevy_theme.rs:374 — unwrap_or(default) should use
  unwrap_or_default(). Replaced.

Build verification:
  cargo fmt --all -- --check                  exit 0
  audit_color32_literals.py                    0 new literals
  audit_bevy_color_literals.py                 0 new literals
  cargo clippy --all-targets --all-features
      -- -D warnings                           clean (40s cold)
  cargo test --lib                             1119 passed / 0 failed

The CI gates that were previously failing on PR #250 (cargo fmt +
color audit) now all pass, AND the cargo clippy step — which had
been silently skipped on the first run — now also passes.

+++ src/ui/construction/cards.rs
@@ -481,7 +481,7 @@ pub fn build_constructed_card_data(
             verb: "Produces",
             amount: super::data::format_power(total_mw),
             per_unit_mw: total_mw,
-            multiplier: count.max(1) as u32,
+            multiplier: count.max(1),
             spare_mw: None,
             insufficient: false,
             tooltip_lines: vec![
@@ -499,7 +499,7 @@ pub fn build_constructed_card_data(
             verb: "Power",
             amount: "0 W".to_string(),
             per_unit_mw: 0.0,
-            multiplier: count.max(1) as u32,
+            multiplier: count.max(1),
             spare_mw: None,
             insufficient: false,
             tooltip_lines: vec!["No grid interaction".to_string()],
@@ -511,7 +511,7 @@ pub fn build_constructed_card_data(
             verb: "Demand",
             amount: super::data::format_power(total_mw),
             per_unit_mw: -total_mw,
-            multiplier: count.max(1) as u32,
+            multiplier: count.max(1),
             spare_mw: None,
             insufficient: false,
             tooltip_lines: vec![
